Added category to maintenance request table.
Category although it is needed for every maintenance request, it is
currently set to be optional in the DB for backwards compatibility.
In the future there will be a migration plan where existing data
will be backfilled to have a category, and once everything is set
to guarantee category will be filled for new requests, we can set
a presence constraint to make it required.